Update DAC testing app
