+++ sdk/bare/usr/dac_test/app_dac_test.c
@@ -1,0 +1,7 @@
+#include "app_dac_test.h"
+#include "task_dac_test.h"
+
+void app_dac_test_init(void)
+{
+	task_dac_test_init();
+}
