+++ sdk/bare/usr/dac_test/app_dac_test.h
@@ -1,0 +1,6 @@
+#ifndef APP_DAC_TEST_H
+#define APP_DAC_TEST_H
+
+void app_dac_test_init(void);
+
+#endif // APP_DAC_TEST_H
